+++ json/src/test/java/org/apache/metamodel/json/JsonDataContextTest.java
@@ -24,7 +24,9 @@ import java.util.Arrays;
 import junit.framework.TestCase;
 
 import org.apache.metamodel.data.DataSet;
+import org.apache.metamodel.query.FunctionType;
 import org.apache.metamodel.schema.Column;
+import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.schema.builder.SchemaBuilder;
 import org.apache.metamodel.schema.builder.SimpleTableDefSchemaBuilder;
@@ -90,7 +92,47 @@ public class JsonDataContextTest extends TestCase {
             assertTrue(ds.next());
             assertEquals("Row[values=[John, Doe, MALE, football, null, null]]", ds.getRow().toString());
             assertTrue(ds.next());
-            assertEquals("Row[values=[John, Doe, MALE, {type=sport, name=soccer}, sport, soccer]]", ds.getRow().toString());
+            assertEquals("Row[values=[John, Doe, MALE, {type=sport, name=soccer}, sport, soccer]]", ds.getRow()
+                    .toString());
+            assertFalse(ds.next());
+        } finally {
+            ds.close();
+        }
+    }
+
+    public void testUseMapValueFunctionToGetFromNestedMap() throws Exception {
+        final Resource resource = new FileResource("src/test/resources/nested_fields.json");
+        final JsonDataContext dataContext = new JsonDataContext(resource);
+
+        final Schema schema = dataContext.getDefaultSchema();
+        assertEquals("[nested_fields.json]", Arrays.toString(schema.getTableNames()));
+
+        final DataSet ds = dataContext.query().from(schema.getTable(0))
+                .select(FunctionType.MAP_VALUE, "name", new Object[] { "first" }).execute();
+        try {
+            assertTrue(ds.next());
+            assertEquals("Row[values=[John]]", ds.getRow().toString());
+            assertTrue(ds.next());
+            assertEquals("Row[values=[John]]", ds.getRow().toString());
+            assertFalse(ds.next());
+        } finally {
+            ds.close();
+        }
+    }
+
+    public void testUseDotNotationToGetFromNestedMap() throws Exception {
+        final Resource resource = new FileResource("src/test/resources/nested_fields.json");
+        final JsonDataContext dataContext = new JsonDataContext(resource);
+
+        final Schema schema = dataContext.getDefaultSchema();
+        assertEquals("[nested_fields.json]", Arrays.toString(schema.getTableNames()));
+
+        final DataSet ds = dataContext.query().from(schema.getTable(0)).select("name.first").execute();
+        try {
+            assertTrue(ds.next());
+            assertEquals("Row[values=[John]]", ds.getRow().toString());
+            assertTrue(ds.next());
+            assertEquals("Row[values=[John]]", ds.getRow().toString());
             assertFalse(ds.next());
         } finally {
             ds.close();